Added assertions to the dimensions of the input to the neural net.
Previously if you passed an input with width or height 0, TensorFlow fromPixels would error because the requested size would be an invalid texture. This wasn't intuitive to the user of the API. These asserts help with catching the error earlier and present it to the end user in a better fashion.