PoseNet - model with segmentation and part map outputs.
This commit adds support for loading and decoding a model with
segmentation and part map outputs, and some new demos showing how they
can be visualized.

This is a WIP.

It adds a new api method to estimate a single pose and get its
segmentation mask and colored part map image.

Currently the segmentation mask and part map image are tensors, but we
may want to figure out how to make these returns tensorless.

Also - do we want the api to returned the colored part map image, or
just the part map?

It temporarily has the demos reference the posenet lib by relative path
so that when developing them they can get live updates.