Fix coco-ssd package name and demo version
The output variable name in the rollup config file should be changed
from coco-ssd to cocoSsd to both avoid dashes in variable names and also
to match the readme sample usage.

This pr also fixes the version listed for coco-ssd in the package.json
and also switches the demo to import the npm package.