converted segmentation model to have tensorless api; it now returns arrays for segmentation and partMap.
For the partMap, it now returns an array with part ids instead of rgb
colors.  If there is no part id in a pixel, that pixel value will have
-1.

Created an op for converting a partMap to an rgb image.

Added utility methods to posenet for masking an image, and applying
a color map to an image based on its part map.