Expose body-pix tensor based ops. (#275)
In many of my personal projects I find myself wanting to work with the tensors returned from bodypix, instead of the cpu based output.    This is particularly the case when wanting to do shader based ops on the outputs.

This PR exposes tensor based ops on the `BodyPix` class.

The two new API methods are:
`estimatePersonSegmentationActivation`
`estimatePartSegmentationOpActivation`

They both take 3d tensors as inputs and return 2d tensors.

The existing API methods are unchanged.  They wrap the new api methods and convert the output to be cpu based arrays.