iOS safari shader compilation issue
It appears that sinFactor can be output into the shader as an integer (-1 or 1).  On iOS (13.4.1) safari this causes a shader compilation issue.  Every other browser I've tried has dealt with it just fine.

I'm not sure this is the best fix but I can confirm it does resolve the issue in iOS (13.4.1) safari.

Below is the relevant parts of the safari console log:

[Log] ERROR: 0:159: '*' : wrong operand types - no operation '*' exists that takes a left-hand operand of type 'highp float' and a right operand of type 'const int' (or there is no acceptable conversion)

int coordX = int(float(x - 424) * 6.123233995736766e-17 - float(y - 188) * -1);